refactor conda and spack configuration classes into their own modules for external use
